requirements.txt
absl-py==1.4.0
amqp==5.3.1
asgiref==3.8.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
Automat==24.8.1
billiard==4.2.1
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.1.1
constantly==23.10.4
cryptography==44.0.1
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-redis==6.0.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
dm-tree==0.1.8
dp-accounting==0.4.3
drf-yasg==1.21.10
et_xmlfile==2.0.0
filelock==3.17.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
joblib==1.4.2
keras==3.8.0
kombu==5.5.4
libclang==18.1.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
numpy==1.26.4
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
packaging==22.0
pandas==2.3.1
parsel==1.10.0
prompt_toolkit==3.0.51
Protego==0.4.0
protobuf==4.25.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
queuelib==1.7.0
redis==6.4.0
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.22.1
tensorflow_privacy==0.9.0
termcolor==2.5.0
threadpoolctl==3.5.0
tldextract==5.1.3
tomli==2.2.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
vine==5.1.0
w3lib==2.3.1
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
zope.interface==7.2
